Updates (#79)

+++ Scripts/Mechanics/Waypoints.cs
@@ -1,8 +1,6 @@
-using System;
 using Server.Network;
 using Server.Engines.Quests;
 using Server.Mobiles;
-using System.Collections.Generic;
 using Server.Engines.PartySystem;
 using System.Linq;
 
@@ -181,21 +179,18 @@ namespace Server
         {
             EnsureCapacity(21 + (name.Length * 2));
 
-            m_Stream.Write((int)serial);
-
+            m_Stream.Write(serial);
             m_Stream.Write((ushort)x);
             m_Stream.Write((ushort)y);
             m_Stream.Write((sbyte)z);
             m_Stream.Write((byte)mapID); //map 
-
             m_Stream.Write((ushort)type);
-
             m_Stream.Write((ushort)(ignoreObject ? 1 : 0));
 
             if(type == WaypointType.Corpse)
-                m_Stream.Write((int)1046414);
+                m_Stream.Write(1046414);
             else
-                m_Stream.Write((int)1062613);
+                m_Stream.Write(1062613);
 
             m_Stream.WriteLittleUniNull(name);
 
@@ -208,7 +203,7 @@ namespace Server
         public RemoveWaypoint(Serial serial)
             : base(0xE6, 5)
         {
-            m_Stream.Write((int)serial);
+            m_Stream.Write(serial);
         }
     }
 }
